OW-453 review link between valuations, trades and portfolios

+++ valuation-app/src/main/java/com/acuo/valuation/providers/acuo/PricingResultPersister.java
@@ -1,7 +1,7 @@
 package com.acuo.valuation.providers.acuo;
 
-import com.acuo.persist.entity.Portfolio;
 import com.acuo.persist.entity.Trade;
+import com.acuo.persist.entity.TradeValuation;
 import com.acuo.persist.entity.Valuation;
 import com.acuo.persist.ids.PortfolioId;
 import com.acuo.persist.services.PortfolioService;
@@ -18,7 +18,6 @@ import lombok.extern.slf4j.Slf4j;
 import javax.inject.Inject;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.TreeSet;
@@ -28,15 +27,13 @@ public class PricingResultPersister implements ResultPersister<PricingResults>, 
 
     private final TradeService<Trade> tradeService;
     private final ValuationService valuationService;
-    private final PortfolioService portfolioService;
     private final ValueService valueService;
     private MarkitValuationProcessor.PricingResultProcessor nextProcessor;
 
     @Inject
-    public PricingResultPersister(TradeService<Trade> tradeService, ValuationService valuationService, PortfolioService portfolioService, ValueService valueService) {
+    public PricingResultPersister(TradeService<Trade> tradeService, ValuationService valuationService, ValueService valueService) {
         this.tradeService = tradeService;
         this.valuationService = valuationService;
-        this.portfolioService = portfolioService;
         this.valueService = valueService;
     }
 
@@ -75,12 +72,7 @@ public class PricingResultPersister implements ResultPersister<PricingResults>, 
         for (Result<MarkitValuation> result : results) {
 
             for (Value value : result.getValue().getValues()) {
-
-
                 String tradeId = value.getTradeId();
-
-                log.debug("tradeId:" + tradeId);
-
                 Trade trade = tradeService.findById(tradeId);
 
                 if (trade == null) {
@@ -94,27 +86,25 @@ public class PricingResultPersister implements ResultPersister<PricingResults>, 
                     for (Valuation valuation : valuations) {
                         if (valuation.getDate().equals(date)) {
                             //existing date, add or replace the value
-                            if (valuation.getValues() == null)
-                                valuation.setValues(new HashSet<>());
-                            Set<com.acuo.persist.entity.Value> existedValues = valuation.getValues();
 
-                            for (com.acuo.persist.entity.Value existedValue : existedValues) {
-                                if (existedValue.getCurrency().equals(currency) && existedValue.getSource().equalsIgnoreCase("Markit")) {
-                                    log.debug("deleting value id [{}]", existedValue.getId());
-                                    try {
-                                        valueService.delete(existedValue.getId());
-                                        existedValues.remove(existedValue);
-                                    } catch (Exception e) {
+                            Set<com.acuo.persist.entity.Value> existedValues = valuation.getValues();
+                            if (existedValues != null) {
+                                for (com.acuo.persist.entity.Value existedValue : existedValues) {
+                                    if (existedValue.getCurrency().equals(currency) && existedValue.getSource().equalsIgnoreCase("Markit")) {
+                                        log.debug("deleting value id [{}]", existedValue.getId());
+                                        try {
+                                            valueService.delete(existedValue.getId());
+                                            existedValues.remove(existedValue);
+                                        } catch (Exception e) {
+                                        }
                                     }
                                 }
                             }
 
                             com.acuo.persist.entity.Value newValue = createValue(currency, value.getPv(), "Markit");
-                            valuation.getValues().add(newValue);
-                            valuationService.createOrUpdate(valuation);
-                            trade.getValuations().add(valuation);
+                            newValue.setValuation(valuation);
+                            valueService.createOrUpdate(newValue);
                             portfolioIds.add(PortfolioId.fromString(trade.getPortfolio().getPortfolioId()));
-                            addsumValuationOfPortfolio(trade.getPortfolio(), date, currency, "Markit", value.getPv());
                             found = true;
                             break;
                         }
@@ -124,32 +114,23 @@ public class PricingResultPersister implements ResultPersister<PricingResults>, 
                 if (!found) {
                     //new valutaion
 
-                    Valuation valuation = createValuation(date, trade.getTradeId());
+                    TradeValuation valuation = createValuation(date, trade.getTradeId());
                     com.acuo.persist.entity.Value newValue = createValue(currency, value.getPv(), "Markit");
-                    Set<com.acuo.persist.entity.Value> values = new HashSet<>();
-                    values.add(newValue);
-                    valuation.setValues(values);
+                    newValue.setValuation(valuation);
 
-                    if (trade.getValuations() != null)
-                        trade.getValuations().add(valuation);
-                    else {
-                        Set<Valuation> valuationSet = new HashSet<Valuation>();
-                        trade.setValuations(valuationSet);
-                    }
-
+                    valuation.setTrade(trade);
+                    valuation.setPortfolio(trade.getPortfolio());
 
                     valuationService.createOrUpdate(valuation);
-                    tradeService.save(trade);
                     portfolioIds.add(PortfolioId.fromString(trade.getPortfolio().getPortfolioId()));
-                    addsumValuationOfPortfolio(trade.getPortfolio(), date, currency, "Markit", value.getPv());
                 }
             }
         }
         return portfolioIds;
     }
 
-    private Valuation createValuation(LocalDate date, String tradeId) {
-        Valuation valuation = new Valuation();
+    private TradeValuation createValuation(LocalDate date, String tradeId) {
+        TradeValuation valuation = new TradeValuation();
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy/MM/dd");
         String dateFormatted = date.format(formatter);
         valuation.setValuationId(dateFormatted + "-" + tradeId);
@@ -163,53 +144,5 @@ public class PricingResultPersister implements ResultPersister<PricingResults>, 
         newValue.setCurrency(currency);
         newValue.setPv(pv);
         return newValue;
-    }
-
-    private void addsumValuationOfPortfolio(Portfolio portfolio, LocalDate date, com.opengamma.strata.basics.currency.Currency currency, String source, Double pv) {
-
-        portfolio = portfolioService.findById(portfolio.getPortfolioId(), 2);
-
-        Valuation theValuation = null;
-        com.acuo.persist.entity.Value theValue = null;
-
-
-        if (portfolio.getValuations() != null) {
-            for (Valuation valuation : portfolio.getValuations()) {
-                if (valuation.getDate().format(DateTimeFormatter.ofPattern("yyyy/MM/dd")).equals(date.format(DateTimeFormatter.ofPattern("yyyy/MM/dd")))) {
-                    theValuation = valuation;
-                    if (valuation.getValues() != null)
-                        for (com.acuo.persist.entity.Value value : valuation.getValues()) {
-                            if (value.getCurrency().equals(currency) && value.getSource().equals(source))
-                                theValue = value;
-                        }
-                }
-            }
-        }
-
-
-        if (theValuation == null) {
-            theValuation = createValuation(date, portfolio.getPortfolioId());
-            theValue = createValue(currency, pv, "Markit");
-            Set<com.acuo.persist.entity.Value> values = new HashSet<>();
-            values.add(theValue);
-            theValuation.setValues(values);
-
-            if (portfolio.getValuations() == null)
-                portfolio.setValuations(new HashSet<Valuation>());
-
-            portfolio.getValuations().add(theValuation);
-            portfolioService.createOrUpdate(portfolio);
-        } else {
-            if (theValue == null) {
-                theValue = createValue(currency, pv, source);
-                if (theValuation.getValues() == null)
-                    theValuation.setValues(new HashSet<>());
-                theValuation.getValues().add(theValue);
-                valuationService.createOrUpdate(theValuation);
-            } else {
-                theValue.setPv(theValue.getPv() + pv);
-                valueService.createOrUpdate(theValue);
-            }
-        }
     }
 }
